Make inputfield a subcomponent of userform
